+++ modules/core/src/component/dynamic-form-control.component.ts
@@ -13,11 +13,13 @@ export abstract class DynamicFormControlComponent implements OnInit {
     customTemplate: TemplateRef<any>;
     hasFocus: boolean;
     model: DynamicFormControlModel;
-    type: string; // must be defined by subclass
 
     incompatibilities: Array<string> = [];
 
-    constructor() {}
+    abstract readonly type: string;
+
+    constructor() {
+    }
 
     ngOnInit() {
 
@@ -58,13 +60,13 @@ export abstract class DynamicFormControlComponent implements OnInit {
         return this.control.valid;
     }
 
-    disable() {
+    disable(): void {
 
         this.control.disable();
         this.model.disabled = true;
     }
 
-    enable() {
+    enable(): void {
 
         this.control.enable();
         this.model.disabled = false;