+++ modules/core/src/model/dynamic-option-control.model.ts
@@ -6,14 +6,14 @@ export interface DynamicFormOptionConfig {
 
     disabled?: boolean;
     label?: string;
-    value: any;
+    value: boolean | number | string;
 }
 
-export class DynamicFormOption implements DynamicFormOptionConfig {
+export class DynamicFormOption {
 
     disabled: boolean;
-    label: string;
-    value: any;
+    label: string | null;
+    value: boolean | number | string;
 
     constructor(config: DynamicFormOptionConfig) {