+++ modules/core/src/model/form-array/dynamic-form-array.model.ts
@@ -1,17 +1,17 @@
 import {ValidatorFn, AsyncValidatorFn} from "@angular/forms";
 import {DynamicFormControlModel, DynamicFormControlModelConfig, ClsConfig} from "../dynamic-form-control.model";
 import {DynamicFormValueControlModel} from "../dynamic-form-value-control.model";
-import {getValue, isFunction, isDefined} from "../../utils";
+import {getValue, isFunction} from "../../utils";
 
 export class DynamicFormArrayGroupModel {
 
     group: Array<DynamicFormValueControlModel<any>>;
-    index: number;
+    index: number | null;
 
-    constructor(group: Array<DynamicFormValueControlModel<any>> = [], index?: number) {
+    constructor(group: Array<DynamicFormValueControlModel<any>> = [], index: number | null = null) {
 
         this.group = group;
-        this.index = isDefined(index) ? index : null;
+        this.index = index;
     }
 }
 
@@ -28,11 +28,13 @@ export interface DynamicFormArrayModelConfig extends DynamicFormControlModelConf
 
 export class DynamicFormArrayModel extends DynamicFormControlModel {
 
-    asyncValidator: AsyncValidatorFn;
+    asyncValidator: AsyncValidatorFn | null;
     createGroup: () => Array<DynamicFormValueControlModel<any>>;
     groups: Array<DynamicFormArrayGroupModel> = [];
     initialCount: number;
-    validator: ValidatorFn;
+    validator: ValidatorFn | null;
+
+    readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_ARRAY;
 
     private originGroup: Array<DynamicFormValueControlModel<any>>; // only to reinstantiate from JSON
 
@@ -46,15 +48,18 @@ export class DynamicFormArrayModel extends DynamicFormControlModel {
 
         this.asyncValidator = getValue(config, "asyncValidator", null);
         this.createGroup = config["createGroup"];
-        this.groups = getValue(config, "groups", null);
         this.initialCount = getValue(config, "initialCount", 1);
         this.originGroup = this.createGroup();
-        this.type = DYNAMIC_FORM_CONTROL_TYPE_ARRAY;
         this.validator = getValue(config, "validator", null);
 
-        if (!Array.isArray(this.groups)) {
+        if (Array.isArray(config.groups)) {
 
-            this.groups = [];
+            config.groups.forEach((arrayGroup, index) => {
+                this.groups.push(new DynamicFormArrayGroupModel(arrayGroup.group, arrayGroup.index || index));
+            });
+
+        } else {
+
             for (let i = 0; i < this.initialCount; i += 1) {
                 this.addGroup();
             }