+++ modules/core/src/model/radio/dynamic-radio-group.model.ts
@@ -7,18 +7,19 @@ export const DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP = "RADIO_GROUP";
 
 export interface DynamicRadioGroupModelConfig extends DynamicOptionControlModelConfig {
 
-    legend?: string;
+    legend?: string | null;
 }
 
 export class DynamicRadioGroupModel<T> extends DynamicOptionControlModel<T> implements DynamicFieldSet {
 
-    legend: string;
+    legend: string | null;
+
+    readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP;
 
     constructor(config: DynamicRadioGroupModelConfig, cls?: ClsConfig) {
 
         super(config, cls);
 
         this.legend = getValue(config, "legend", null);
-        this.type = DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP;
     }
 }