+++ modules/core/src/model/select/dynamic-select.model.ts
@@ -5,10 +5,10 @@ export const DYNAMIC_FORM_CONTROL_TYPE_SELECT = "SELECT";
 
 export class DynamicSelectModel<T> extends DynamicOptionControlModel<T> {
 
+    readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_SELECT;
+
     constructor(config: DynamicOptionControlModelConfig, cls?: ClsConfig) {
 
         super(config, cls);
-
-        this.type = DYNAMIC_FORM_CONTROL_TYPE_SELECT;
     }
 }