+++ modules/core/src/service/dynamic-form.service.ts
@@ -8,7 +8,10 @@ import {
     DYNAMIC_FORM_CONTROL_TYPE_ARRAY
 } from "../model/form-array/dynamic-form-array.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_GROUP, DynamicFormGroupModel} from "../model/form-group/dynamic-form-group.model";
-import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP, DynamicCheckboxGroupModel} from "../model/checkbox/dynamic-checkbox-group.model";
+import {
+    DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP,
+    DynamicCheckboxGroupModel
+} from "../model/checkbox/dynamic-checkbox-group.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX, DynamicCheckboxModel,} from "../model/checkbox/dynamic-checkbox.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_INPUT, DynamicInputModel} from "../model/input/dynamic-input.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP, DynamicRadioGroupModel} from "../model/radio/dynamic-radio-group.model";
@@ -18,11 +21,7 @@ import {DYNAMIC_FORM_CONTROL_TYPE_TEXTAREA, DynamicTextAreaModel} from "../model
 @Injectable()
 export class DynamicFormService {
 
-    formBuilder: FormBuilder;
-
-    constructor(formBuilder: FormBuilder) {
-
-        this.formBuilder = formBuilder;
+    constructor(private formBuilder: FormBuilder) {
     }
 
     createFormArray(model: DynamicFormArrayModel): FormArray {
@@ -36,7 +35,7 @@ export class DynamicFormService {
         return this.formBuilder.array(formArray, model.validator, model.asyncValidator);
     }
 
-    createFormGroup(group: Array<DynamicFormControlModel>, groupExtra: {[key: string]: any} = null): FormGroup {
+    createFormGroup(group: Array<DynamicFormControlModel>, groupExtra: {[key: string]: any} | null = null): FormGroup {
 
         let formGroup = {};
 
@@ -102,7 +101,7 @@ export class DynamicFormService {
         return group.find(controlModel => controlModel.id === id);
     }
 
-    fromJSON(json: Array<Object>): Array<DynamicFormControlModel> {
+    fromJSON(json: Array<Object>): Array<DynamicFormControlModel> | never {
 
         let formModel: Array<DynamicFormControlModel> = [];
 
@@ -146,7 +145,7 @@ export class DynamicFormService {
                     break;
 
                 default:
-                    throw new Error(`no known form control type defined on JSON object`);
+                    throw new Error(`unknown form control type defined on JSON object`);
             }
         });