+++ modules/core/src/utils.ts
@@ -1,33 +1,8 @@
-// Object.assign Polyfill 
-if (typeof Object.assign !== "function") {
-    (function () {
-        Object.assign = function (target) {
-            "use strict";
-            if (target === undefined || target === null) {
-                throw new TypeError("Cannot convert undefined or null to object");
-            }
-
-            var output = Object(target);
-            for (var index = 1; index < arguments.length; index++) {
-                var source = arguments[index];
-                if (source !== undefined && source !== null) {
-                    for (var nextKey in source) {
-                        if (source.hasOwnProperty(nextKey)) {
-                            output[nextKey] = source[nextKey];
-                        }
-                    }
-                }
-            }
-            return output;
-        };
-    })();
-}
-
 export function isDefined(object: any): boolean {
     return object !== undefined && object !== null;
 }
 
-export function isEmptyString(_string: string): boolean {
+export function isEmptyString(_string: string | null | undefined): boolean {
     return typeof _string !== "string" || _string.length === 0;
 }