+++ modules/ui-material/src/dynamic-form-material.component.ts
@@ -31,7 +31,8 @@ export class DynamicFormMaterialComponent extends DynamicFormControlComponent {
     @ViewChild(MdRadioGroup) mdRadioGroup: MdRadioGroup;
 
     incompatibilities: Array<string> = [DYNAMIC_FORM_CONTROL_TYPE_SELECT, DYNAMIC_FORM_CONTROL_TYPE_TEXTAREA];
-    type: string = DYNAMIC_FORM_UI_MATERIAL;
+
+    readonly type: string = DYNAMIC_FORM_UI_MATERIAL;
 
     constructor() {
         super();