+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# RC.1
+
+### **Breaking Changes** 
+
+* **!!! IMPORTANT !!!** UMD bundles now provided in `bundles` sub folder 
+
+### **Bugfixes** 
+
+* UMD bundles now work correctly for all module systems (closes [#106](https://github.com/udos86/ng2-dynamic-forms/issues/106)) 
+
+### **Features**
+
+* Upgrade to TypeScript 2.0 (closes [#108](https://github.com/udos86/ng2-dynamic-forms/issues/108))  
+* Minified UMD bundles added to npm packages (closes [#105](https://github.com/udos86/ng2-dynamic-forms/issues/105)) 
+
+
 # beta.17
 
 ### **Breaking Changes**